+++ System_Files/system.app.launcher.external.local.app.name.konsole.terminal.php
@@ -10,9 +10,9 @@
 // Source Code. It makes it easier, instead of the User to manually change the directory
 // to the default directory by putting 'cd', which does to the default home directory, 
 // unless the default directory has been specified by the administrator or by the user's 
-// preferences. Please see the "system.app.launcher.external.app.name"
+// preferences. Please see the "system.app.launcher.external.app.name.konsole.terminal.sh"
 
-$launch = shell_exec('cd');
+$launch = shell_exec('sh system.app.launcher.external.local.app.name.konsole.terminal.sh');
 echo "<pre>$launch</pre>";
 
 ?>